Canpolicy_model_building/requirements.txt
langchain-community
llama-index
llama-index-llms-huggingface-api
llama-index-embeddings-langchain
python-dotenv
qdrant_client
llama-index-vector-stores-qdrant